+++ src/ULinkRPC.Starter/StarterModels.cs
@@ -16,7 +16,35 @@ internal enum SerializerKind
 internal enum ClientEngineKind
 {
     Unity,
+    Tuanjie,
     Godot
+}
+
+internal static class ClientEngineKindExtensions
+{
+    public static bool IsUnityCompatible(this ClientEngineKind clientEngine) => clientEngine switch
+    {
+        ClientEngineKind.Unity => true,
+        ClientEngineKind.Tuanjie => true,
+        ClientEngineKind.Godot => false,
+        _ => throw new ArgumentOutOfRangeException(nameof(clientEngine), clientEngine, null)
+    };
+
+    public static string GetDisplayName(this ClientEngineKind clientEngine) => clientEngine switch
+    {
+        ClientEngineKind.Unity => "Unity",
+        ClientEngineKind.Tuanjie => "Tuanjie",
+        ClientEngineKind.Godot => "Godot",
+        _ => throw new ArgumentOutOfRangeException(nameof(clientEngine), clientEngine, null)
+    };
+
+    public static string GetStarterClientLabel(this ClientEngineKind clientEngine) => clientEngine switch
+    {
+        ClientEngineKind.Unity => "Unity 2022 LTS",
+        ClientEngineKind.Tuanjie => "Tuanjie (Unity-compatible)",
+        ClientEngineKind.Godot => "Godot 4.6",
+        _ => throw new ArgumentOutOfRangeException(nameof(clientEngine), clientEngine, null)
+    };
 }
 
 internal sealed record ResolvedVersions(